Update sketch.js
final design：On the original artwork animation, I iterated the white circle according to the audio interaction, so that the color and animation of the white circle changed. The current effect is that you can see the brightness and size of the green circle change with the effect of the audio. This is imagined to look like a streetlight in an urban streetscape.